Resolver refactor (#181)
* Fixed the override key handling.

* Fixed docs and print issues.

* Initial branch commit WIP

* Working on getting all the unittests working again.

* Finished fixing unittests.

* Fixed style issue.
