+++ lib/pavilion/test_config/variables.py
@@ -173,10 +173,7 @@ permutation.
 
         # Create a new var set manager for each permutation.
         for perm in permutations:
-            var_man = VariableSetManager()
-
-            var_man.variable_sets = copy.deepcopy(self.variable_sets)
-            var_man.deferred = self.deferred
+            var_man = copy.deepcopy(self)
 
             for (var_set, var), idx in perm.items():
                 new_list = [self.variable_sets[var_set][var][idx]]
@@ -605,6 +602,16 @@ index, sub_var) tuple.
                     .format(list(def_parsed.keys()))
                 )
 
+    def __deepcopy__(self, memodict=None):
+        """Deeply copy this variable set manager."""
+
+        var_man = VariableSetManager()
+
+        var_man.variable_sets = copy.deepcopy(self.variable_sets)
+        var_man.deferred = self.deferred
+
+        return var_man
+
     def __contains__(self, item):
 
         var_set, var, index, sub_var = self.parse_key(item)
@@ -806,8 +813,8 @@ dict with a single None: value pair."""
         for key, value in value_pairs.items():
             if not isinstance(value, str):
                 raise VariableError(
-                    "Variable values must be unicode strings, got '{}'"
-                    .format(type(value)), sub_var=key)
+                    "Variable values must be unicode strings, got '{}' as '{}'"
+                    .format(value, type(value)), sub_var=key)
 
             self.data[key] = value
