+++ lib/pavilion/unittest.py
@@ -14,12 +14,12 @@ from pathlib import Path
 
 from pavilion import arguments
 from pavilion import config
-from pavilion import pav_vars
+from pavilion import pavilion_variables
 from pavilion import system_variables
 from pavilion.test_run import TestRun
 from pavilion.test_config.file_format import TestConfigLoader
 from pavilion.test_config import VariableSetManager
-from pavilion.test_config import resolve_config
+from pavilion.test_config import resolver
 from pavilion.output import dbg_print
 
 
@@ -102,7 +102,7 @@ base class.
 
         self.pav_cfg.pav_cfg_file = cfg_path
 
-        self.pav_cfg.pav_vars = pav_vars.PavVars()
+        self.pav_cfg.pav_vars = pavilion_variables.PavVars()
 
         if not self.pav_cfg.working_dir.exists():
             self.pav_cfg.working_dir.mkdir(parents=True)
@@ -313,6 +313,7 @@ The default config is: ::
         return cfg
 
     __config_lines = pprint.pformat(QUICK_TEST_BASE_CFG).split('\n')
+    # Code analysis indicating format isn't found for 'bytes' is a Pycharm bug.
     _quick_test_cfg.__doc__ = _quick_test_cfg.__doc__.format(
         '\n'.join(['    ' + line for line in __config_lines]))
     del __config_lines
@@ -346,7 +347,7 @@ The default config is: ::
         if sched_vars is not None:
             var_man.add_var_set('sched', sched_vars)
 
-        cfg = resolve_config(cfg, var_man, [])
+        cfg = resolver.TestConfigResolver.resolve_config(cfg, var_man)
 
         test = TestRun(
             pav_cfg=self.pav_cfg,
@@ -381,9 +382,11 @@ class ColorResult(unittest.TextTestResult):
 
     def __init__(self, *args, **kwargs):
         self.stream = None
+        self.showAll = None
         super().__init__(*args, **kwargs)
 
     def startTest(self, test):
+        """Write out the test description (with shading)."""
         super().startTest(test)
         if self.showAll:
             self.stream.write(self.GREY)
@@ -393,21 +396,25 @@ class ColorResult(unittest.TextTestResult):
             self.stream.flush()
 
     def addSuccess(self, test):
+        """Write the success text in green."""
         self.stream.write(self.GREEN)
         super().addSuccess(test)
         self.stream.write(self.COLOR_RESET)
 
     def addFailure(self, test, err):
+        """Write the Failures in magenta."""
         self.stream.write(self.MAGENTA)
         super().addFailure(test, err)
         self.stream.write(self.COLOR_RESET)
 
     def addError(self, test, err):
+        """Write errors in red."""
         self.stream.write(self.RED)
         super().addError(test, err)
         self.stream.write(self.COLOR_RESET)
 
     def addSkip(self, test, reason):
+        """Note skips in cyan."""
         self.stream.write(self.CYAN)
         super().addSkip(test, reason)
         self.stream.write(self.COLOR_RESET)
