+++ test/data/pav_config_dir/hosts/layer_host.yaml
@@ -1,2 +1,1 @@
-slurm:
-  partition: host
+summary: host
