+++ test/data/pav_config_dir/modes/layer_mode.yaml
@@ -1,2 +1,1 @@
-slurm:
-  partition: mode
+summary: mode
