+++ test/tests/resolver_tests.py
@@ -1,30 +1,38 @@
+"""Test the various components of the test resolver."""
 
 import copy
 
 from pavilion import plugins
 from pavilion import system_variables
-from pavilion.test_config import (load_test_configs,
-                                  TestConfigError,
-                                  apply_overrides,
-                                  resolve_permutations,
-                                  resolve_config)
+from pavilion.test_config import TestConfigError, resolver
 from pavilion.test_config import variables, string_parser
 from pavilion.unittest import PavTestCase
 
 
-class TestSetupTests(PavTestCase):
+class ResolverTests(PavTestCase):
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+
+    def setUp(self):
+        """Initialize plugins and setup a resolver."""
+        plugins.initialize_plugins(self.pav_cfg)
+
+        self.resolver = resolver.TestConfigResolver(self.pav_cfg)
+
+    def tearDown(self):
+        """Reset the plugins."""
+        plugins._reset_plugins()
 
     def test_loading_tests(self):
         """Make sure get_tests can find tests and resolve inheritance."""
 
-        plugins.initialize_plugins(self.pav_cfg)
-
-        tests = load_test_configs(self.pav_cfg, 'this', [], ['hello_world'])
+        tests = self.resolver.load(['hello_world'], host='this')
         self.assertEqual(sorted(['narf', 'hello', 'world']),
-                         sorted([test['name'] for test in tests]))
+                         sorted([test['name'] for test, _ in tests]))
 
-        tests = load_test_configs(self.pav_cfg, 'this', [], ['hello_world.hello'])
-        hello = tests.pop()
+        tests = self.resolver.load(['hello_world.hello'], host='this')
+        hello, _ = tests.pop()
 
         # There should have only been 1
         self.assertFalse(tests)
@@ -34,41 +42,32 @@ class TestSetupTests(PavTestCase):
         # Make sure the variable from the host config got propagated.
         self.assertIn('hosty', hello['variables'])
 
-        tests = load_test_configs(self.pav_cfg, 'this', [], ['hello_world.narf'])
-        narf = tests.pop()
+        tests = self.resolver.load(['hello_world.narf'], host='this')
+        narf, _ = tests.pop()
         # Make sure this got overridden from 'world'
         self.assertEqual(narf['scheduler'], 'dummy')
         # Make sure this didn't get lost.
         self.assertEqual(narf['run']['cmds'], ['echo "Running World"'])
 
-        plugins._reset_plugins()
-
     def test_loading_hidden(self):
         """Make sure we only get hidden tests when specifically requested."""
 
-        plugins.initialize_plugins(self.pav_cfg)
-
-        tests = load_test_configs(self.pav_cfg, 'this', [], ['hidden'])
-        names = sorted([t['name'] for t in tests])
+        tests = self.resolver.load(['hidden'], 'this', [], {})
+        names = sorted([t['name'] for t, _ in tests])
         self.assertEqual(names, ['hello', 'narf'])
 
-        tests = load_test_configs(self.pav_cfg, 'this', [],
-                                  ['hidden._hidden'])
-        names = sorted([t['name'] for t in tests])
+        tests = self.resolver.load(['hidden._hidden'], 'this', [], {})
+        names = sorted([t['name'] for t, _ in tests])
         self.assertEqual(names, ['_hidden'])
-
-        plugins._reset_plugins()
 
     def test_layering(self):
         """Make sure test config layering works as expected."""
-
-        plugins.initialize_plugins(self.pav_cfg)
 
         for host in ('this', 'layer_host'):
             for modes in ([], ['layer_mode']):
                 for test in ('layer_tests.layer_test',
                              'layer_tests.layer_test_part'):
-                    answer = 'standard'
+                    answer = ''
                     if host == 'layer_host':
                         answer = 'host'
                     if modes:
@@ -76,56 +75,45 @@ class TestSetupTests(PavTestCase):
                     if test.endswith('part'):
                         answer = 'test'
 
-                    tests = load_test_configs(
-                        pav_cfg=self.pav_cfg,
+                    tests = self.resolver.load(
+                        [test],
                         host=host,
-                        modes=modes,
-                        tests=[test])
-                    self.assertEqual(tests[0]['slurm']['partition'], answer)
-
-        plugins._reset_plugins()
+                        modes=modes)
+                    test, _ = tests[0]
+                    self.assertEqual(test['summary'], answer)
 
     def test_defaulted_variables(self):
         """Make sure default variables work as expected."""
 
-        plugins.initialize_plugins(self.pav_cfg)
-
-        tests = load_test_configs(
-            self.pav_cfg,
+        tests = self.resolver.load(
+            tests=['defaulted.test'],
             host='defaulted',
             modes=['defaulted'],
-            tests=['defaulted.test']
         )
 
-        test = tests[0]
+        test, _ = tests[0]
         self.assertEqual(test['variables']['host_def'], ['host'])
         self.assertEqual(test['variables']['mode_def'], ['mode'])
         self.assertEqual(test['variables']['test_def'], ['test'])
         self.assertNotIn('no_val', test['variables'])
 
         with self.assertRaises(TestConfigError):
-            tests = load_test_configs(
-                self.pav_cfg,
+            self.resolver.load(
+                tests=['defaulted_error.error'],
                 host='defaulted',
                 modes=['defaulted'],
-                tests=['defaulted_error.error']
             )
-
-        plugins._reset_plugins()
 
     def test_extended_variables(self):
         """Make sure default variables work as expected."""
 
-        plugins.initialize_plugins(self.pav_cfg)
-
-        tests = load_test_configs(
-            self.pav_cfg,
+        tests = self.resolver.load(
+            tests=['extended.test'],
             host='extended',
             modes=['extended'],
-            tests=['extended.test']
         )
 
-        test = tests[0]
+        test, _ = tests[0]
         self.assertEqual(test['variables']['long_base'],
                          ['what', 'are', 'you', 'up', 'to', 'punk?'])
         self.assertEqual(test['variables']['single_base'],
@@ -139,33 +127,36 @@ class TestSetupTests(PavTestCase):
         self.assertEqual(test['variables']['null_base'],
                          ['test'])
 
-
-        plugins._reset_plugins()
-
     def test_apply_overrides(self):
         """Make sure overrides get applied to test configs correctly."""
-
-        plugins.initialize_plugins(self.pav_cfg)
-
-        cfgs = load_test_configs(self.pav_cfg, 'this', [], ['hello_world'])
 
         overrides = [
             'slurm.num_nodes=3',
             'run.cmds.0="echo nope"',
         ]
 
-        bad_key_overrides = [
+        bad_overrides = [
+            # No such variables
             'florp.blorp=3',
+            # You can't override the scheduler
             'scheduler="raw"',
+            # run.cmds is a list
             'run.cmds.eeeh="hello"',
+            # Invalid index.
             'run.cmds.10000="ok"',
+            # Summary isn't a dict.
             'summary.nope=blarg',
-            '""="empty"'
+            # A very empty key.
+            '""="empty"',
+            # Value is an incomplete mapping.
+            "summary={asdf",
         ]
 
-        for cfg in cfgs:
+        cfgs = self.resolver.load(['hello_world'], 'this',
+                                  overrides=overrides)
+
+        for cfg, _ in cfgs:
             alt_cfg = copy.deepcopy(cfg)
-            apply_overrides(alt_cfg, overrides)
 
             # Make sure the overrides were applied
             self.assertEqual(alt_cfg['slurm']['num_nodes'], "3")
@@ -175,19 +166,10 @@ class TestSetupTests(PavTestCase):
             self.assertEqual(cfg['run']['env'], alt_cfg['run']['env'])
 
         # Make sure we get appropriate errors in several bad key cases.
-        for bad_key_over in bad_key_overrides:
-            for cfg in cfgs:
-                alt_cfg = copy.deepcopy(cfg)
-                with self.assertRaises(KeyError):
-                    apply_overrides(alt_cfg, [bad_key_over])
-
-        # Check for appropriate value errors too.
-        for cfg in cfgs:
-            alt_cfg = copy.deepcopy(cfg)
-            with self.assertRaises(ValueError):
-                apply_overrides(alt_cfg, ["summary={asdf"])
-
-        plugins._reset_plugins()
+        for bad_override in bad_overrides:
+            with self.assertRaises(TestConfigError):
+                self.resolver.load(['hello_world'], 'this',
+                                   overrides=[bad_override])
 
     def test_resolve_permutations(self):
         """Make sure permutations are applied correctly."""
@@ -214,8 +196,12 @@ class TestSetupTests(PavTestCase):
         }
 
         orig_permutations = raw_test['variables']
+        var_man = copy.deepcopy(self.resolver.base_var_man)
+        var_man.add_var_set('var', raw_test['variables'])
 
-        test, permuted = resolve_permutations(raw_test, {}, {})
+        test, permuted = self.resolver.resolve_permutations(
+            raw_test, var_man
+        )
 
         # Foo should triple the permutations, bar should double them -> 6
         # baz shouldn't have an effect (on the permutation count).
@@ -242,14 +228,19 @@ class TestSetupTests(PavTestCase):
             self.assertIn(comb_dict, combinations)
 
     def test_deferred_errors(self):
+        """Using deferred variables in inappropriate places should raise
+        errors."""
         test = {
             'permute_on': ['sys.def'],
             'variables': {},
         }
 
+        var_man = variables.VariableSetManager()
+        var_man.add_var_set('sys', {'def': variables.DeferredVariable()})
+
         with self.assertRaises(TestConfigError):
-            resolve_permutations(
-                test, {}, {'def': variables.DeferredVariable()})
+            self.resolver.resolve_permutations(
+                test, var_man)
 
         test = {
             'permute_on': ['foo.1'],
@@ -259,7 +250,7 @@ class TestSetupTests(PavTestCase):
         }
 
         with self.assertRaises(TestConfigError):
-            resolve_permutations(test, {}, {})
+            self.resolver.resolve_permutations(test, var_man)
 
         test = {
             'permute_on': ['no_exist'],
@@ -267,7 +258,7 @@ class TestSetupTests(PavTestCase):
         }
 
         with self.assertRaises(TestConfigError):
-            resolve_permutations(test, {}, {})
+            self.resolver.resolve_permutations(test, var_man)
 
     def test_resolve_vars_in_vars(self):
         test = {
@@ -279,7 +270,11 @@ class TestSetupTests(PavTestCase):
             }
         }
 
-        test, permuted = resolve_permutations(test, {}, {'soda':'pepsi'})
+        var_man = variables.VariableSetManager()
+        var_man.add_var_set('sys', {'soda': 'pepsi'})
+        var_man.add_var_set('var', test['variables'])
+
+        test, permuted = self.resolver.resolve_permutations(test, var_man)
 
         possible_stuff = ['yapple-apple-x',
                           'yapple-pepsi',
@@ -305,12 +300,13 @@ class TestSetupTests(PavTestCase):
             'permute_on': []
         }
 
+        var_man = copy.deepcopy(self.resolver.base_var_man)
+        var_man.add_var_set('var', test['variables'])
+
         with self.assertRaises(variables.VariableError):
-            resolve_permutations(test, {}, {})
+            self.resolver.resolve_permutations(test, var_man)
 
     def test_finalize(self):
-
-        plugins.initialize_plugins(self.pav_cfg)
 
         cfg = self._quick_test_cfg()
 
@@ -343,12 +339,10 @@ class TestSetupTests(PavTestCase):
         results = test.gather_results(test.run())
         test.save_results(results)
 
-        plugins._reset_plugins()
-
     def test_resolve_all_vars(self):
+        """Most of the variable resolution stuff is tested elsewhere,
+        but it's good to have it all put together in one final test."""
 
-        # Most of the variable resolution stuff is tested elsewhere,
-        # but it's good to have it all put together in one final test.
         test = {
             'build': {
                 'cmds':
@@ -385,23 +379,24 @@ class TestSetupTests(PavTestCase):
         }
 
         # This is all that changes between the two.
-        import copy
         answer2 = copy.deepcopy(answer1)
         answer2['build']['cmds'] = ["echo 2 4", "echo 2", "echo 4a"]
 
         answers = [answer1, answer2]
 
-        test, permuted = resolve_permutations(test,
-                                              # Pav vars
-                                              {'nope': '9'},
-                                              # sys vars
-                                              {'nope': '10'})
+        var_man = variables.VariableSetManager()
+        var_man.add_var_set('pav', {'nope': '9'})
+        var_man.add_var_set('sys', {'nope': '10'})
+        var_man.add_var_set('var', test['variables'])
+        del test['variables']
+
+        test, permuted = self.resolver.resolve_permutations(test, var_man)
 
         self.assertEqual(len(permuted), 2)
 
         # Make sure each of our permuted results is in the list of answers.
         for var_man in permuted:
-            out_test = resolve_config(test, var_man, ['build'])
+            out_test = self.resolver.resolve_config(test, var_man)
             self.assertIn(out_test, answers)
 
         # Make sure we can successfully disallow deferred variables in a
@@ -414,21 +409,23 @@ class TestSetupTests(PavTestCase):
             'variables': {}
         }
 
-        dvar = variables.DeferredVariable()
+        var_man = variables.VariableSetManager()
+        var_man.add_var_set('sys', {'foo': variables.DeferredVariable()})
+        var_man.add_var_set('var', {})
 
-        test, permuted = resolve_permutations(test, {}, {'foo': dvar})
+        test, permuted = self.resolver.resolve_permutations(test, var_man)
 
         with self.assertRaises(string_parser.ResolveError):
             # No deferred variables in the build section.
-            resolve_config(test, permuted[0], ['build'])
+            self.resolver.resolve_config(test, permuted[0])
 
     def test_env_order(self):
         """Make sure environment variables keep their order from the test
         config to the final run scripts."""
 
-        plugins.initialize_plugins(self.pav_cfg)
+        test_confs = self.resolver.load(['order'], host='this')
 
-        test_conf = load_test_configs(self.pav_cfg, 'this', [], ['order'])[0]
+        test_conf, _ = test_confs[0]
 
         test = self._quick_test(test_conf, "order")
 
@@ -459,5 +456,3 @@ class TestSetupTests(PavTestCase):
                              "Got the following instead: \n{}"
                              .format(''.join(["{}: {}\n".format(*v) for v in
                                               exports])))
-
-        plugins._reset_plugins()
